+++ lib/multi_json/adapter_selector.rb
@@ -142,7 +142,7 @@ module MultiJson
       when ::Symbol then load_adapter_by_name(adapter_spec.to_s)
       when nil, false then load_adapter(default_adapter)
       when ::Module then adapter_spec
-      else raise ::LoadError, adapter_spec
+      else raise ::LoadError, "expected adapter to be a Symbol, String, or Module, got #{adapter_spec.inspect}"
       end
     rescue ::LoadError => e
       raise AdapterError.build(e)